+++ the_problem_manager/lib/model/datesModel.dart
@@ -1,4 +1,4 @@
-class DatesData {
+class DatesModel {
   String start;
   String end;
 }
