+++ the_problem_manager/lib/model/nameModel.dart
@@ -1,4 +1,4 @@
-class NameData {
+class NameModel {
   String yourName = "";
   String personName = "";
   bool samePerson = false;
