+++ the_problem_manager/pubspec.yaml
@@ -24,10 +24,16 @@ dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+
+  sqflite: ^2.0.0+3
+  path_provider: ^2.0.1
+  intl: 0.16.1
+
+  bloc: 7.0.0
+  flutter_bloc: 7.0.0
 
 dev_dependencies:
   flutter_test:
